+++ java-manta-client-unshaded/src/main/java/com/joyent/manta/http/StandardHttpHelper.java
@@ -762,7 +762,7 @@ public class StandardHttpHelper implements HttpHelper {
     }
 
     @Override
-    public void close() throws Exception {
+    public void close() throws IOException {
         connectionContext.close();
     }
 }